Added erlang.js which defines erlang terms
Updated gulp dist task so that the order of files is correct.

Refactored the CLI module.

ElixirScript takes the input given to it and makes it into one
big es6 module. The es6 imports and the module name are taken
from a new config file (either exjs.exs by default).