Disable "no-console" linting for specific console.warns
These all are exceptional cases of warnings where the warning is
far more important than adhering to a style guide on lints, even if
some of them might potentially show up in the browser. Each is a
hand-picked exception, so these use in-file comment overrides.

Rationales are included in preceding comments.